requirements.txt
bleach==6.2.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.59.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
regex==2025.9.1
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
python-telegram-bot==20.7
gunicorn==21.2.0
